Dev (#78)
* Fixed the focus on the table

So that it won't move to the new row every time there is a new row

* Fixed a bug in view profile

Bug: view profile button was not shown unless number of profiles > 1

* Fixed #71

The new auto mapper allows the user to discover new mappings inside the repo

* Fixed visual for showing repo item

If the short name of a repo item is empty, the namespace will be shown

* Fixed visual bug in mapping

The mapping was not shown when the process name is empty. It is now fixed

* Added removal of files in the repo

* Fixed #76

* #74 Added import from plain text

There is dependency on other API to support the function, and development for the full support will take some time. A simple import from text function is implemented for now.

* Added 3D model display

* Small changes

Change the icon for mapping
Let the search function search in links

* Fixed visual problem

Gateway was not aligned properly

* Fixed #77

* Fixed a bug in import

* Fixed a typo on metadata

* Fixed a bug in map presentation

* Added a temp solution for reading STS XML

* Fixed the default value problem in Table Class view

* Added auto translate

Translate text (SMART documents) to models (MMEL)

* Fixed a bug in element import